refactor: (31.03.24) Re-export `MessageEntity` from `Message` module
