+++ client/src/Message/index.ts
@@ -1,2 +1,3 @@
 export * from "./components";
 export * from "./stores";
+export * from "./types/MessageEntity";
