+++ client/src/Report/stores/ReportedMessagesStore.ts
@@ -1,7 +1,7 @@
 import {mergeWith} from "lodash";
 import {AbstractEntityStore} from "../../entity-store";
 import {EntitiesPatch} from "../../entities-store";
-import {MessageEntity} from "../../Message/types";
+import {MessageEntity} from "../../Message";
 import {convertMessageToNormalizedForm} from "../../Message/utils";
 import {Message} from "../../api/types/response";
 import {mergeCustomizer} from "../../utils/object-utils";
